+++ sandbox/grommet-app/src/pages/sticker-sheet/content/Controls/index.js
@@ -1,0 +1,6 @@
+export * from './Accordions';
+export * from './Buttons';
+export * from './Menus';
+export * from './Paginations';
+export * from './Tabs';
+export * from './ToggleGroups';
