+++ sandbox/grommet-app/src/pages/sticker-sheet/content/index.js
@@ -1,1 +1,2 @@
+export * from './Controls';
 export * from './Type';
